// src/data/storyCritters.gen.ts
export interface CritterArt {
  src: number;
  frames: number;
  /** One frame's pixel size in the strip (aspect ratio source). */
  w: number;
  h: number;
}

export const CRITTER_ART = {
  fox: { src: require('../../assets/story/critters/fox.webp'), frames: 12, w: 373, h: 200 },
  fox_idle: { src: require('../../assets/story/critters/fox_idle.webp'), frames: 12, w: 373, h: 200 },
  deer: { src: require('../../assets/story/critters/deer.webp'), frames: 12, w: 230, h: 200 },
  deer_idle: { src: require('../../assets/story/critters/deer_idle.webp'), frames: 12, w: 230, h: 200 },
  koi: { src: require('../../assets/story/critters/koi.webp'), frames: 12, w: 178, h: 200 },
  alpaca: { src: require('../../assets/story/critters/alpaca.webp'), frames: 12, w: 214, h: 200 },
  alpaca_idle: { src: require('../../assets/story/critters/alpaca_idle.webp'), frames: 12, w: 214, h: 200 },
  snake_sand: { src: require('../../assets/story/critters/snake_sand.webp'), frames: 12, w: 206, h: 200 },
  snake_sand_idle: { src: require('../../assets/story/critters/snake_sand_idle.webp'), frames: 12, w: 206, h: 200 },
  trex: { src: require('../../assets/story/critters/trex.webp'), frames: 12, w: 306, h: 200 },
  trex_idle: { src: require('../../assets/story/critters/trex_idle.webp'), frames: 12, w: 306, h: 200 },
  triceratops: { src: require('../../assets/story/critters/triceratops.webp'), frames: 12, w: 334, h: 200 },
  triceratops_idle: { src: require('../../assets/story/critters/triceratops_idle.webp'), frames: 12, w: 334, h: 200 },
  raptor: { src: require('../../assets/story/critters/raptor.webp'), frames: 10, w: 284, h: 200 },
  raptor_idle: { src: require('../../assets/story/critters/raptor_idle.webp'), frames: 12, w: 284, h: 200 },
  husky: { src: require('../../assets/story/critters/husky.webp'), frames: 12, w: 235, h: 200 },
  husky_idle: { src: require('../../assets/story/critters/husky_idle.webp'), frames: 12, w: 235, h: 200 },
  arctic_wolf: { src: require('../../assets/story/critters/arctic_wolf.webp'), frames: 12, w: 360, h: 200 },
  arctic_wolf_idle: { src: require('../../assets/story/critters/arctic_wolf_idle.webp'), frames: 12, w: 360, h: 200 },
  stag: { src: require('../../assets/story/critters/stag.webp'), frames: 12, w: 181, h: 200 },
  stag_idle: { src: require('../../assets/story/critters/stag_idle.webp'), frames: 12, w: 181, h: 200 },
  frog: { src: require('../../assets/story/critters/frog.webp'), frames: 12, w: 324, h: 200 },
  frog_idle: { src: require('../../assets/story/critters/frog_idle.webp'), frames: 12, w: 324, h: 200 },
  snake: { src: require('../../assets/story/critters/snake.webp'), frames: 12, w: 206, h: 200 },
  snake_idle: { src: require('../../assets/story/critters/snake_idle.webp'), frames: 12, w: 206, h: 200 },
  wasp: { src: require('../../assets/story/critters/wasp.webp'), frames: 8, w: 186, h: 200 },
  piranha: { src: require('../../assets/story/critters/piranha.webp'), frames: 12, w: 227, h: 200 },
  dolphin: { src: require('../../assets/story/critters/dolphin.webp'), frames: 12, w: 319, h: 200 },
  whale: { src: require('../../assets/story/critters/whale.webp'), frames: 12, w: 283, h: 200 },
  manta: { src: require('../../assets/story/critters/manta.webp'), frames: 12, w: 289, h: 200 },
  clownfish: { src: require('../../assets/story/critters/clownfish.webp'), frames: 12, w: 252, h: 200 },
  buffalo: { src: require('../../assets/story/critters/buffalo.webp'), frames: 12, w: 282, h: 200 },
  buffalo_idle: { src: require('../../assets/story/critters/buffalo_idle.webp'), frames: 12, w: 282, h: 200 },
  horse: { src: require('../../assets/story/critters/horse.webp'), frames: 12, w: 232, h: 200 },
  horse_idle: { src: require('../../assets/story/critters/horse_idle.webp'), frames: 12, w: 232, h: 200 },
  donkey: { src: require('../../assets/story/critters/donkey.webp'), frames: 12, w: 223, h: 200 },
  donkey_idle: { src: require('../../assets/story/critters/donkey_idle.webp'), frames: 12, w: 223, h: 200 },
  astro_frog: { src: require('../../assets/story/critters/astro_frog.webp'), frames: 12, w: 128, h: 200 },
  astro_bee: { src: require('../../assets/story/critters/astro_bee.webp'), frames: 12, w: 127, h: 200 },
  saucer: { src: require('../../assets/story/critters/saucer.webp'), frames: 1, w: 254, h: 200 },
} satisfies Record<string, CritterArt>;

export type CritterKey = keyof typeof CRITTER_ART;